Adjustments to issues when compiling in 64 bits

+++ engine/server/application/LoginServer/src/shared/SessionApiClient.cpp
@@ -120,7 +120,7 @@ void SessionApiClient::OnSessionValidate(const apiTrackingNumber trackingNumber,
 	else
 	{
 		// attempt to validate a CS Tool associated with this request.
-		CSToolConnection::validateCSTool( ( uint32 ) userData, result, session );
+		CSToolConnection::validateCSTool( ( uintptr_t ) userData, result, session );
 	}
 
 }
